% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwp_function.R
\name{kwp}
\alias{kwp}
\title{kwp() function}
\usage{
kwp(file_path = NULL, xvar = NULL, yvar = NULL)
}
\arguments{
\item{xvar}{the name of the column in the dataset that contains the group names.}

\item{yvar}{the name of the column in the dataset that contains the values to compare between groups.}
}
\value{
If there is a significant difference between groups, a table with pairwise comparisons and p-values will be printed. Otherwise, a message suggesting to try another dataset will be printed.
}
\description{
This function allows you to Perform Kruskal-Wallis test; it shows
"~~No significance between groups, try with other dataset~~"
if there is no significant between groups in the dataset.
Otherwise it will show
~~Observe Significance between groups, continue pairwise_wilcox comparisons~~
and calculate pairwise comparisons between groups, showing p value.
}
\details{
it will check dplyr package; if hasn't installed it, it will install automatically.

Run kwp() and the read file window will prompt out, please locate the dataset which
value in the 2nd column and groups in 3rd column.

There are some datasets in the data folder; please try them.
function code goes here
}
\examples{
kwp()

}
