% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt1_function.R
\name{ppt1}
\alias{ppt1}
\title{ppt1() function}
\description{
This function allows you to create a special plot; it shows
the data distribution in a line of a group.
You can easy to compare different groups the data distribution in lines.
}
\details{
It will check ggplot2 package; if hasn't installed it, it will install automatically.

The read file window will prompt out, please locate the dataset which
values in the 2nd column and groups in 3rd column.

There are some datasets in the dataset folder; please try them.
}
