% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt2_function.R
\name{ppt2}
\alias{ppt2}
\title{ppt2() function}
\description{
This function allows you to create a box plot; A box plot displays the median
,quartiles, and outliers of each group's data.You can easy to compare different groups.
}
\details{
It will check ggpubr package; if hasn't installed it, it will install automatically.

The read file window will prompt out, please locate the dataset which
values in the 2nd column and groups in 3rd column.

There are some datasets in the dataset folder; please try them.
}
