% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tampa_temp_data_documentation.R
\docType{data}
\name{tampa_temp_data}
\alias{tampa_temp_data}
\title{Tampa Monthly Temperature Data}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\usage{
data(tampa_temp_data)
}
\value{
A data frame object containing the Tampa monthly temperature data.
}
\description{
This data set contains monthly temperature data for Tampa from 1900 to 2022.
}
\examples{
data(tampa_temp_data)
}
\keyword{dataset}
