% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tampa_temp_data_2017_2022_documentation.R
\docType{data}
\name{tampa_temp_data_2017_2022}
\alias{tampa_temp_data_2017_2022}
\alias{tampa_temp_data}
\title{Tampa Monthly Temperature Data 2017-2022}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\usage{
data(tampa_temp_data_2017_2022)
}
\value{
A data frame object containing the Tampa monthly temperature data 2017-2022
}
\description{
This data set contains monthly temperature data for Tampa from 2017 to 2022
}
\examples{
data(tampa_temp_data_2017_2022)
}
\keyword{dataset}
